chore: add exclusive gating features for owned and shared contexts
